fix: failing test for blog like functionality (fixes #1)
The locator to search for the likes is updated to use a `data-testid`.
This required to move the likes text element inside `<span></span>`.

With logging, it is observed that it takes more than 3 seconds to render
the new likes. Thus, the timeout for the test is increased to 5000ms.
This ensures that the like button click operation has taken effect.

Note that this is an eventual consistency made possible by using
expect().toPass()

Lastly, unnecessary waits have been removed and a single likes
element is referenced throughout.