fix(redux-bloglist): send user info in POST response (fixes #3)
This provides a better solution to modify the response of creating a
new blog. No need to change the response data in the frontend.

Note that a new schema rule has been defined for `toObject` to remove
the `_id` and `__v` fields. This is done since Mongoose preserves the
original document structure during save. With these changes, the
backend provides the newly created blog object that is similar to the
modifications performed by the frontend earlier.