feat(query-bloglist): create a notification custom hook to use across components
Multiple components need notification capability. The way it worked
until now was to show notification for a fixed number of seconds and
then hide it.

This functionality is now wrapped in a custom hook where notification
messages are dispatched and hidden after 5 seconds. There is already a
NotificationContextProvider used throughout App therefore, this custom
hook works seamlessly across components.